Remade part of the code for final demo
Each of the file contains part of each subsystem code. In addition,
structs are utilized to make the code look simpler and easier to
implement. Only the partplacerstate.ino and the supportfunctionsStructs
are currently implemented with tray and part stepper active. the magnets
and servos for the part placer are also implemented